Hardcode (incorrect) divisor for ORION
While the ORION token contract is in fact registred with a multiplier/divisor of 100M the pylon contracts are doing some magic
when returning pending rewards. Since the FE is using this divisor to convert the pending rewards into "human readable form"

We incorrect divisor here, otherwise rewards are mis-represented.