feat: announce new tx with created at so it can be used for efficient query
Since we're storing our data in a timeseries database, providing
datetimes for the query can significantly speed up the query. It helps
the underlying datastore to pick the right partition/chunk/index to
locate the data.

Also added an index on `id` to avoid having to do table scans, but
rather index scans in the correct partition.